Emit the problem-stats and summary blocks from the solver core
The Ipopt-style stats and end-of-run summary were printed by the CLI, so the
Python and C frontends never saw them. Move the console printers from
pounce-cli to the shared pounce-solve-report::console crate, and have
pounce-algorithm's solve() emit the problem-statistics block (before the
iteration table) and the summary (after), gated on print_level. Every
frontend now gets the identical log at print_level >= 1; print_level 0 is
silent. JSON mode forces print_level 0, so structured output is unaffected.

Add IpoptNlp::eval_counts() (default zeros; OrigIpoptNlp reports its live
counters) and drain it into SolveStatistics, fulfilling those fields'
long-standing "used to populate SolveStatistics" intent, so the summary's
evaluation tallies come from the solver's own tracking rather than a
frontend counting wrapper. Expose pounce.print_banner() for in-process
frontends (the banner stays frontend-emitted, before the problem is read).

CLI stdout is byte-identical on the NLP fixtures except the summary's
obj/grad/constr/jac evaluation counts, which now report the solver's true
count (one lower than the old CLI wrapper, which counted a frontend-side
evaluation), and print_level 0 now also suppresses the stats/summary blocks.
