+++ Cargo.toml
@@ -55,7 +55,7 @@ default-members = [
 ]
 
 [workspace.package]
-version = "0.8.0"
+version = "0.9.0"
 edition = "2024"
 license = "EPL-2.0"
 authors = ["John Kitchin <jkitchin@andrew.cmu.edu>"]
@@ -66,23 +66,23 @@ repository = "https://github.com/jkitchin/pounce"
 # deps lack a registry version). Crates pick these up via
 # `pounce-foo.workspace = true`.
 [workspace.dependencies]
-pounce-common = { path = "crates/pounce-common", version = "0.8.0" }
-pounce-linalg = { path = "crates/pounce-linalg", version = "0.8.0" }
-pounce-linsol = { path = "crates/pounce-linsol", version = "0.8.0" }
-pounce-nlp = { path = "crates/pounce-nlp", version = "0.8.0" }
-pounce-nl = { path = "crates/pounce-nl", version = "0.8.0" }
-pounce-hsl = { path = "crates/pounce-hsl", version = "0.8.0" }
-pounce-feral = { path = "crates/pounce-feral", version = "0.8.0" }
-pounce-algorithm = { path = "crates/pounce-algorithm", version = "0.8.0" }
-pounce-restoration = { path = "crates/pounce-restoration", version = "0.8.0" }
-pounce-presolve = { path = "crates/pounce-presolve", version = "0.8.0" }
-pounce-l1penalty = { path = "crates/pounce-l1penalty", version = "0.8.0" }
-pounce-qp = { path = "crates/pounce-qp", version = "0.8.0" }
-pounce-convex = { path = "crates/pounce-convex", version = "0.8.0" }
-pounce-sensitivity = { path = "crates/pounce-sensitivity", version = "0.8.0" }
-pounce-solve-report = { path = "crates/pounce-solve-report", version = "0.8.0" }
-pounce-studio-core = { path = "crates/pounce-studio-core", version = "0.8.0" }
-pounce-observability = { path = "crates/pounce-observability", version = "0.8.0" }
+pounce-common = { path = "crates/pounce-common", version = "0.9.0" }
+pounce-linalg = { path = "crates/pounce-linalg", version = "0.9.0" }
+pounce-linsol = { path = "crates/pounce-linsol", version = "0.9.0" }
+pounce-nlp = { path = "crates/pounce-nlp", version = "0.9.0" }
+pounce-nl = { path = "crates/pounce-nl", version = "0.9.0" }
+pounce-hsl = { path = "crates/pounce-hsl", version = "0.9.0" }
+pounce-feral = { path = "crates/pounce-feral", version = "0.9.0" }
+pounce-algorithm = { path = "crates/pounce-algorithm", version = "0.9.0" }
+pounce-restoration = { path = "crates/pounce-restoration", version = "0.9.0" }
+pounce-presolve = { path = "crates/pounce-presolve", version = "0.9.0" }
+pounce-l1penalty = { path = "crates/pounce-l1penalty", version = "0.9.0" }
+pounce-qp = { path = "crates/pounce-qp", version = "0.9.0" }
+pounce-convex = { path = "crates/pounce-convex", version = "0.9.0" }
+pounce-sensitivity = { path = "crates/pounce-sensitivity", version = "0.9.0" }
+pounce-solve-report = { path = "crates/pounce-solve-report", version = "0.9.0" }
+pounce-studio-core = { path = "crates/pounce-studio-core", version = "0.9.0" }
+pounce-observability = { path = "crates/pounce-observability", version = "0.9.0" }
 # feral: pinned to the published crates.io release so the pounce crates stay
 # publishable — a git rev would block the crates.io publish and is rejected by
 # scripts/check-release-consistency.sh ("git dependency … needs a crates.io
