+++ crates/pounce-solve-report/Cargo.toml
@@ -16,6 +16,10 @@ pounce-linsol.workspace = true
 pounce-nlp = { workspace = true, features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+# Logo/banner ANSI styling for the console printers (moved here with the
+# printers from pounce-cli, #206).
+anstream.workspace = true
+anstyle.workspace = true
 
 [lints]
 workspace = true
